requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.6.0
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.41.1
Pillow==10.4.0
litellm==1.52.0
pydantic-ai==1.21.0
pydantic-ai-slim==1.21.0
google-genai==1.52.0
celery==5.4.0
redis==5.0.1
unidecode==1.3.8
numpy==1.26.4
requests>=2.32.3,<3.0.0
pgvector==0.3.4
beautifulsoup4==4.12.3
lxml==5.3.0
nest-asyncio==1.6.0

